requirements.txt
PyYAML>=6.0
python-dateutil>=2.8
pytz>=2024.1
requests>=2.31
beautifulsoup4>=4.12
python-dotenv>=1.0
# Optional: LLM for the discovery agent (free tier)
google-generativeai>=0.7
